feature: integrate settings to pkg/ functionalities
- Integrated settings to editor related functionalities.
- Moved open with vi function to utils.go and renamed as
  open-via-editor.
