+++ pkg/utils_test.go
@@ -9,6 +9,7 @@ import (
 	"os"
 	"testing"
 
+	"github.com/anonistas/notya/lib/models"
 	"github.com/anonistas/notya/pkg"
 )
 
@@ -278,5 +279,51 @@ func TestListDir(t *testing.T) {
 		})
 
 		td.deletingFunc(td.folderName)
+	}
+}
+
+func TestOpenViaEditor(t *testing.T) {
+	type utilArgs struct {
+		filename       string
+		stdargs        models.StdArgs
+		settings       models.Settings
+		deleteFileFunc func(filename string)
+		createFileFunc func(filename string)
+	}
+
+	tests := []struct {
+		testName string
+		ua       utilArgs
+		expected error
+	}{
+		{
+			testName: "should open created exiting file properly",
+			ua: utilArgs{
+				filename: "test_file.txt",
+				stdargs:  models.StdArgs{},
+				settings: models.InitSettings(),
+				deleteFileFunc: func(filename string) {
+					pkg.Delete(filename)
+				},
+				createFileFunc: func(filename string) {
+					pkg.WriteNote(filename, []byte{})
+				},
+			},
+
+			expected: errors.New("exit status 1"),
+		},
+	}
+
+	for _, td := range tests {
+		t.Run(td.testName, func(t *testing.T) {
+			td.ua.createFileFunc(td.ua.filename)
+
+			got := pkg.OpenViaEditor(td.ua.filename, td.ua.stdargs, td.ua.settings)
+			if got != td.expected && got.Error() != td.expected.Error() {
+				t.Errorf("Sum was different, Got: %v | Want: %v", got, td.expected)
+			}
+
+			td.ua.deleteFileFunc(td.ua.filename)
+		})
 	}
 }
